Use packet loss in lagging sender and saver

+++ cotta-core/src/main/kotlin/com/mgtriffid/games/cotta/core/config/CottaConfig.kt
@@ -39,7 +39,7 @@ interface DebugConfig {
 
             interface Issues {
                 val latency: Latency
-                val packetLoss: Long
+                val packetLoss: Double
             }
         }
     }
